Quota page: Display AZ capacity for instance type
Show number of AZs that have capacity for each instance type.

The capacity is colour coded, so that an operator can get a quick
indication of the latency between requesting that instance type, and
the fulfillment of that request.
- Red means only one AZ
- Yellow means more than one, but less than all AZs
- Green means all AZs

Note I have chosen to do this per-instance, instead of per-GPU class,
because eventually there will be a GPU-type that is used across
different instance families, where their availability/capacity is different.